deploy_model/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
Flask-HTTPAuth
Flask-RESTful